Rename dom_helpers back to dom.
